requirements.txt
antlr4-python3-runtime==4.13.0
pytest>=7.4.2
tox>=4.11.3
